allow customise the generation of id
- new interface: transport.IDMaker, the type of id is string
- each kind of task can have its own IDMaker
- the identify of one task is (name, id)

+++ transport/marshaller.go
@@ -1,10 +1,10 @@
 package transport
 
 var Encode = struct {
-	Default  int16
-	JSON     int16
-	GOB      int16
-	JSONSAFE int16
+	Default  int
+	JSON     int
+	GOB      int
+	JSONSAFE int
 }{
 	0, 1, 2, 3,
 }
